feat(sales): extract CustomerFields and ProductFields components
- Extract customer fields (RUT, name, date, payment method) to CustomerFields (163 lines)
- Extract product fields (folio) to ProductFields (37 lines)
- SalesFormFields.jsx reduced from 202 to 45 lines (thin orchestrator)
- All files ≤200 lines, zero behavior change
